+++ app/app.vue
@@ -28,7 +28,6 @@ useSeoMeta({
 
 <template>
   <UApp>
-    <NuxtPwaManifest />
     <AppHeader />
     <UMain>
       <NuxtPage />